+++ r-package/R/utils.R
@@ -57,66 +57,63 @@ set_max_street_time <- function(max_walk_dist, walk_speed, max_trip_duration) {
 #' Select transport mode
 #'
 #' @param mode character string passed from routing functions.
+#' @param mode_egress character string passed from routing functions.
 #'
 #' @family support functions
 
-select_mode <- function(mode) {
-
-  mode <- toupper(unique(mode))
+select_mode <- function(mode, mode_egress) {
 
   # list all available modes
   dr_modes  <- c('WALK','BICYCLE','CAR','BICYCLE_RENT','CAR_PARK')
   tr_modes  <- c('TRANSIT', 'TRAM','SUBWAY','RAIL','BUS','FERRY','CABLE_CAR','GONDOLA','FUNICULAR')
   all_modes <- c(tr_modes, dr_modes)
 
-  # check for invalid input
+  # check for invalid input --------------------------------------------------
+  mode <- toupper(unique(mode))
+  mode_egress <- toupper(unique(mode_egress))[1]
+
   lapply(mode, function(x) {
     if (!x %chin% all_modes) {
       stop(paste0(x, " is not a valid 'mode'.\nPlease use one of the following: ",
-                  paste(unique(all_modes), collapse = ", ")))
-    }
-  })
+                  paste(unique(all_modes), collapse = ", "))) }
+    })
+  lapply(mode_egress, function(x) {
+    if (!x %chin% dr_modes) {
+      stop(paste0(x, " is not a valid 'mode'.\nPlease use one of the following: ",
+                  paste(unique(dr_modes), collapse = ", "))) }
+    })
 
-  # assign modes accordingly
+  # assign modes accordingly  --------------------------------------------------
   direct_modes <- mode[which(mode %chin% dr_modes)]
   transit_mode <- mode[which(mode %chin% tr_modes)]
-  if ("TRANSIT" %in% transit_mode) transit_mode <- tr_modes
 
-  # if only a direct_mode is passed, all others are empty
-  if (length(direct_modes) != 0 & length(transit_mode) == 0) {
-    egress_mode <- access_mode <- mode[which(mode %chin% c('WALK', 'BICYCLE'))]
-    if (length(access_mode) == 0){ egress_mode <- access_mode <- 'WALK' }
-    if (direct_modes == "CAR"){ egress_mode <- access_mode <- 'CAR' }
+  # No public transport
+  if ( length(transit_mode) == 0) {
+    if (sum(mode %in% c('WALK', 'BICYCLE')>0)){ direct_modes <- access_mode <- mode[which(mode %chin% c('WALK', 'BICYCLE'))][1] }
+    if (sum(mode %in% "CAR")>0){ direct_modes <- access_mode <- 'CAR' }
     transit_mode <- ""
-  } else
+    mode_egress <- ""
 
-    # if only transit mode is passed, assume 'WALK' as access_ and egress_modes
-    if (length(transit_mode) != 0 & length(direct_modes) == 0) {
-      egress_mode <- access_mode <- mode[which(mode %chin% c('WALK', 'BICYCLE'))]
-      if (length(access_mode) == 0){ egress_mode <- access_mode <- 'WALK' }
-      direct_modes <- ""
+  } else {
 
-    } else
+  # with public transport
+    # all pt modes
+    if ("TRANSIT" %in% transit_mode){ transit_mode <- tr_modes }
 
-      # if transit & direct modes are passed, consider direct as access & egress_modes
-      if (length(transit_mode) != 0 & length(direct_modes) != 0) {
-        access_mode <- direct_modes[which(direct_modes %chin% c('WALK', 'BICYCLE', 'CAR'))]
-        egress_mode <- access_mode <- unique(c('WALK', access_mode))
-      }
+    # if only transit mode is passed, assume 'WALK' as access_mode
+    if (length(direct_modes) == 0) { access_mode <- direct_modes <- 'WALK' }
 
-  # removes CAR and BICYCLE from egress_mode, since a transit user (probably)
-  # won't have to these modes after a transit ride
-  egress_mode <- egress_mode[! egress_mode %chin% c("CAR", "BICYCLE")]
+    # if transit & direct modes are passed, consider direct as access & egress_modes
+    if (length(direct_modes) != 0) { access_mode <- direct_modes }
+  }
 
   # create output as a list
   mode_list <- list('direct_modes' = paste0(direct_modes, collapse = ";"),
                     'transit_mode' = paste0(transit_mode, collapse = ";"),
                     'access_mode'  = paste0(access_mode, collapse = ";"),
-                    'egress_mode'  = paste0(egress_mode, collapse = ";"))
-
+                    'egress_mode'  = paste0(mode_egress[1], collapse = ";"))
 
   return(mode_list)
-
 }