+++ r-package/NEWS.md
@@ -1,7 +1,17 @@
 # log history of r5r package development
 
 -------------------------------------------------------
-# r5r v0.3-0 dev
+# r5r v0.3-1 dev
+
+**Minor changes**
+- Allow for combination of bicycle and public transport. Closed #135.
+- Added new parameter `mode_egress` to routing functions, so that user can  explicitly
+the transport mode used after egress from public transport. Closed #63.
+
+
+
+-------------------------------------------------------
+# r5r v0.3-0
 
 **Major changes**
 - Added Conveyal's R5 repo as a git submodule. This will help improve the long term integration between r5r and R5. Closed #105.
@@ -10,6 +20,8 @@
 **Minor changes**
 - Add columns with population and number of schools in sample data set of Porto Alegre to allow for accessibility examples. Closed #128.
 - the `percentiles` parameter in the `travel_time_matrix` function now only accepts up to 5 cut points due to changes in R5.
+
+
 
 -------------------------------------------------------
 # r5r v0.2-1
@@ -21,6 +33,8 @@ complex/realistic examples.
 sets `time_window` parameter. Closes #127.
 * remove repeated bus route alternatives from the output from `detailed_itineraries`
 * Explicitly link destination points to street network before starting. Closes #121
+
+
 
 -------------------------------------------------------
 # r5r v0.2-0